extracted the logic of funds to its own Component
